requirements.txt
Flask==3.0.3
twilio==8.2.0
gspread==5.7.2
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
openai
python-dotenv == 1.0.1 #currently used version but not necessary to stay same
